pyproject.toml
[project]
name = "voiceMonitor"
version = "0.1.0"
description = "Real-Time Vocal Fatigue Monitoring for Continuous Speech Analytics"
readme = "README.md"
authors = [
  { name="Muhammad Khubaib Ahmad", email="muhammadkhubaibahmad854@gmail.com" }
]
license = { text="MIT" }

requires-python = ">=3.10"
dependencies = [
    "auralis-vfs>=1.1.0",
    "sounddevice>=0.4",
    "pydub>=0.25",
    "soundfile>=0.12",
    "tqdm>=4.65"
]

[project.scripts]
voicemonitor = "voiceMonitor.cli:main"